Implement CAR CID iterator and supplier
Implements mechanisms to list the CIDs given a CAR path, regardless of
the CAR's version. The mechanism allows paths to be put and removed, and
supplies CID iterators given a CAR ID.

Implement mechanism to generate CID IDs for a CAR file, regardless of
its version.

Implement a CID iterator reader that given a CID iterator and a
marshaller converts it into a stream of bytes as an `io.ReadCloser`.

Implement tests using sample CAR files in versions 1 and 2.

Run `gofumpt -l -w .` across the whole repo for consistent formatting.

Leave TODOs on undecided or improvable implementations to contain the PR
scope.

Increase test timeout flaky engine tests.