Register CAR supplier in daemon engine
Instantiate a CAR supplier that uses the same underlying datastore as
the engine, convert it to `CidCallback` and register it. Define a new
flag for daemon that allows configuration of CAR read option on how to
treat zero-length blocks.

Adjust the prefix of datastore entries produced by CAR supplier to avoid
clash since datastore is shared with engine and the supplier.

Add utility to convert a streaming CID supplier to the non-streaming
`CidCallback`. Note, we want to retain the streaming abilities
implemented already for future use, e.g. chunking of CID list etc.

Write tests that assert draining a streaming CID iterator is correct.

Unexport flags until they need to be exported and remove placeholders.