Refactor supplier to work with `LookupKey` interface changes
Change Supplier to work with `LookupKey` instead of CID as a key to
lookup list of CIDs by.

Integrate engine into CAR supplier, such that when a CAR is put/removed
it will notify the engine. It will also self register when instantiated.

Generate mock interfaces for core engine to ease unit testing.

Adjust tests to assert that the engine is indeed notified when CAR put
or remove is called using mock interfaces.

Add CAR supplier placeholder in admin server in preparation for
implementing the CAR `import` command.