Containerize the provider by defining a `Dockerfile`
Add a dockerfile that builds the reference provider with a conditional
initialiazation build argument that is enabled by default. Note, the
docker image built as a result is meant for testing purposes only, since
once initialized it will contain a generated private key that is unsafe
to use in production.