Use iterator instead of double channel in `provider.Callback`
Introduce `MultihashIterator` type to replace the double channel
returned by `provider.Callback`. This is to simplify the usage for
clients.

Require callback to take context in order to gracefully handle the
lifecycle of gorutines that may be started as part of
`MultihashIterator`.

Implement utility class that converts CARv2 `MultihashSortedIndex` into
a `MultihashIterator`. This type is not exported. But may be in the
future if useful to existing clients.

Refactor use of "cid" in names now that provider works with multihashes.