Refactor `Callback` API due to iteration determinism changes
Address TODOs left in code to refactor the `Callback` API now that all
multihashes are read and sorted when CAR multihash iterator is
constructed.

As a result we do not need `context` in CAR mh iterator implementation
of `Callback` anymore since there is no gorutine in the background to
manage the lifecycle of. We do however keep the `context` in signature
of `Callback` for user whom wish to use an alternative implementation of
`Callback`.

Fail fast when `ForEach` traversal of multihashes done in constructor
of CAR multihash iterator returns error. This makes a simpler error
handling logic and makes sense to fail fast since the `Callback`
function already returns error.

Relates to #62