Implement admin server endpoint to remove CARs
Implement a HTTP edpoint exposed on admin server that accepts context ID
associated to a CAR that is already imported and advertises its removal.

Note the endpoint does not implement removal by path alone. This is
because the CAR supplier only accepts put with explicit context ID. If
context ID can be generated as the hash of path, then that functionality
seem to be better housed on the admin CLI side or re-introduced in CAR
supplier itself then added to the admin server.

At the engine level, making the end to end functionality work depends
on:
 - https://github.com/filecoin-project/storetheindex/pull/104

This commit focuses on the admin server interface only.

Fixes #77