Implement `remove car` CLI command
Implement a CLI command that advertises removal of a list of multihashes
previously advertised via importing a CAR file. The command accepts
either the key or the path by which the CAR was imported in order to
identify which CAR file to remove. For the command to work the CAR file
must have previously been advertised.

Note that an optional CAR path flag is added in this commit, since the
existing flag object is mandatory and used in `import car` sub-command.

Relates to:
 - https://github.com/filecoin-project/indexer-reference-provider/pull/83

Fixes #78