Simplify admin server error response
Use a simple `http.Error` when responding with error in admin server.
Reflect changes in the CLI so that the response body is displayed as
expected.

Use consistent way to construct http error in admin server, and reuse
the log message in http response.

Add utility to interact with admin HTTP server in CLI commands to reduce
duplicate code. Refactor CLI commands to use the utilities for both
parsing of error response from admin server and sending of HTTP
requests.

Fixes #80