Do not close the wrapped datastore used for caching
The index provider engine as well as its internal components, like
`dsCache` take an existing datastore instance and reuse it internally.
`dsChace` implements the `datastore.Datastore` interface by wapping the
existing instance. Since the instance is instantiated externally, it
should not be closed off by the provider `Engine` nor `dsChache`.

Instead, delegate the management of the datastore lifecycle to where it
is instantiated, i.e. the `daemon` command.