Remove libp2p server exposed by provider and rewrite e2e tests
Remove the libp2p server in provider in favour of go-legs since it
already provides the functionality of that server.

The sever was used in e2e testing of car data transfer. Rewrite the
tests to use go-legs graphsyc endpoint for fetching advertisements. Tidy
up tests as needed while at it.

Fixes #110