Re-enable flaky tests and remove `time.Sleep` s using event tracer
Now that go-legs is heavily refactored, re-enable the flaky tests and
remove explicit `time.Sleep` calls in the tests. This is so that we can
check if tests remain flaky and if so how to fix them so that they are
not.

Implement a bespoke pubsub event tracer that allows tests to observe and
wait for pubsub events before proceeding with other assertions.

Refactor assertions to reduce duplicate code across engine tests.

Skip ad removal test pending fix for #174.

Fixes #12