Implement a CLI command to verify multihash ingestion
Implement a CLI tool that given a source of multihashes it verifies that
every multihash is indexed by an indexer node and is associated to the
expected peer ID as its provider. The sources of multihash may be CAR
files, CARv2 iterable multihash index files or Lotus markets dagstore
repo (migrated to the latest CAR index format).

This utility is written to aid miners verify their deals are indexed
by an indexer node.

Fixes #137