Remove lotus-specific source in favour of generic api call in verify CLI
Future PR will introduce a graphsync call to use provider API itself as
a source of multihashes. For now remove the lotus-specific command to
keep provider generic.

This change is added in a separate commit to preserve the functionality
implemented for lotus if for whatever reason we want to re-use it.