Implement recursive ingest verification by walking the chain
Add option to `verify-ingest` to walk the chain of advertisement from
either a given CID or head up to a configured limit, and for each
encountered advertisement verify ingest.

Add the ability to limit entries chain sync when it is too long along
with output to tell the CLI user of their existence.