Re implement caching mechanism as `CachedEntriesChunker`
Introduce the `EntriesChunker` interface that given a multihash iterator
would chunk the supplied multihashes into their interlinked `EntryChunk`
IPLD representation.

The chunker includes both chunking and caching mechanism because, the
cache must guarantee that either an entire chunk is cached or none of
it, while maintaining that guarantee for overlapping chains, i.e. chains
that share the same section.

Use the same datastore prefix as the old `lrustore` such that upon
instantiation the previously persisted values, if any, are pruned. Since
eviction did not function and any existing cache is likely larger than
capacity. In addition, the new implementation stores keys to the root
of chains and a counter for repeated links, which are required for the
`CachedEntriesChunker` to work correctly.

Fixes #166, #73