Fix RNG used by `testutil` to avoid intermittent test failures
The `testutil` functions that use RNG, were creating a new instance of
it seeded with current unix time. This causes two main issues: 1)
randomness in tests is not deterministic, which makes reproducing
intermittent test failures difficult. 2) the implementation of current
system nano time is platform-specific and when tests are run
concurrently this can cause issues with duplicate generated values
across gorutines that expect unique values for testing purposes.

Fix this by requiring RNG in `testutil` and instantiate one explicitly
with a fixed seed in each test.