Work around legs subscriber instantiating its own datatrasfer manager
An option to set datatransfer manager for go-legs subscriber was
removed. As a result a libp2p host cannot be used to instantiate
datatransfer manager and go-legs subscriber at the same time.

This approach was used during e2e testing in index-provider. This commit
instantiates a new libp2p host for the subscriber so that we can get
the tests running again and upgrade go-legs dependency until the options
on go-legs side are reviewed.

Relates to:
 - https://github.com/filecoin-project/go-legs/issues/95