Make network announcements optional and introduce engine options
Implement the ability to disable network announcements of published
advertisements. Change the behaviour of engine, so that announcements
are disabled by default until `PublisherKind` is set.

Refactor the engine package, and introduce options with defaults such
that the barrier to instantiate an engine is low with sensible defaults.
Future PRs will refactor the config package to match the options.