New announce command to announce latest advertisement
The `announce` command publishes an announce for the latest advertisement.  This is useful to get indexers to sync without having to publish a new advertisement.

Additional change: On daemon startup, initialize the daemon with the latest ad CID but do not publish it.  Publishing the ad CID is not useful since the daemon will not yet have bootstrapped to any peers.  Having the `announce` command is a better as it allows optional publishing once the daemon is ready.