Smaller default MultihashIterator implementation
The default MultihashIterator implementation does not need to keep the offsets if a check is done for duplicates when the iterator is created. This reduces memory allocated for each iterator instance, by n uint64 values, at the cost of one O(n) iteration of the sorted offsets to check for duplicates.

The default implementation is also exported so that it can be used in other packages that only need a simple multihash iterator.