Do not bother restoring cache when purge cache is set
The existing behaviour will restore the cache regardless of whether the
purge-cache flag is set or not. The rationale for this behaviour was
that the Clear logic needed to know what caches to delete in the
first place.

This rationale no longer stands because of the changes to recovery from
corrupt cache introduced in earlier PRs. The Clear logic would delete
all entries in the datastore regardless to make sure that the datastore
size does not leak as a result of orphan cache entries for example.

The changes here, therefore, won't bother restoring the cache if it is
to be cleared.

Relates to:
 - https://github.com/filecoin-project/index-provider/issues/237