Cache links of a root entry instead of resolving them on the fly
The chunk cache only stored the root of a chunked entry chain. The cache
restore logic then traversed the DAG to resolve all child links. In a
case where the cache is large and the DAG is big it could be slow to
traverse the dag and resolve all the child links.

Instead of resolving them dynamically, store the child links with the
root link so that restoring the hash is fast and won't require
traversing the DAG. The trade-off however is slightly larger storage
consumption, where for each entry chain DAG a mapping of root link to
list of all links in the DAG is stored in the datastore.

Fixes: https://github.com/filecoin-project/index-provider/issues/237