Publish container image to ECR on merge to `main` and release
Add CI job to build and publish the container image to ECR on merges to
`main` and on release with deterministic tagging.

Add a CI job to build the dockerfile and make sure it is functional on
push and PR.

Fix a build issue in dockerfile, and remove the error-prone `init` build
argument. Because, a production environment would not want the docker
image to contain peer identity since it is typically mounted, and a
standalone testing app would benefit from a utility script to set things
up since as it stood the init in dockerfile was broken anyway.