Calculate import cart key from the car file name instead of path
If the import car key is not specified by the user, it is currently
calulated as a hash of the unclean car path.

So different paths to the same file would produce different key and
hence different ad for the very same content.

For example importing:

my.car
./my.car
../my.car
/abs/path/to/my.car

will result in a different ad alhough these paths point to the very same
file.

It would be better if the car file name is used to determine the import
car key.