Mirror an existing provider with optional entries remap and CDN
Implement the ability to periodically check the advertisement chain
published by an index provider an mirror it with optional remapping of
the original entries in HAMT or entry chunk.

Host the original ad chain in the mirror such that the mirror can also
act as a CDN for the original ad chain.

Implement a mirror testing environment that asserts various
configuration permutations function as expected.

Implement utilities to extract `MultihashIterator` from links stored by
an IPLD link system regardless of the underlying entries type.