Add periodic ad flush (#351)
Add periodic ad flush

Currently the last chunk gets stuck in the delegated routing server until it's full. That might be a long time given that re-provide is relatively infrequent. This commit adds a worker that periodically checks the current chunk and publishes it if it's not empty and too old.